Adapts parse_log.py to read more than one log file
This is necessary because we have one log file for each server that runs a
broker. This assumes you have the logs in data/<hostname>_broker.log and saves
data to data/worker-<worker>.csv, appending the results for each log file.
Before it starts running, the old ones are renamed (to <filename>.old). This
saves data from the last run, but overrides data from the ones before this.